requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.2
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.2
cryptography==45.0.6
cyclopts==3.22.5
debugpy==1.8.14
decorator==5.2.1
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastapi-mcp==0.4.0
fastmcp==2.11.3
filelock==3.19.1
filetype==1.2.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
identify==2.6.13
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langsmith==0.4.16
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.13.1
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numpy==2.2.6
openai==1.101.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python-headless==4.12.0.88
orjson==3.11.2
packaging==25.0
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
pre_commit==4.3.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.8.0
pypdfium2==4.30.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
safetensors==0.6.2
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
surya-ocr==0.15.4
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.4
tomli==2.2.1
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
wcwidth==0.2.13
Werkzeug==3.1.1
wheel==0.45.1
zai-sdk==0.0.3.3
zstandard==0.24.0
